'downstream' directory
