+++ ofat.py
@@ -4,11 +4,8 @@ import argparse
 import glob
 import json
 import pandas as pd
-import matplotlib.pyplot as plt
-from pathlib import Path
 
 import validator
-import config
 import utils
 
 """
@@ -123,26 +120,6 @@ def analyse(root_path):
         "tp": tps
     })
 
-def plot(root_path):
-
-    utils.assert_path_exists(root_path)
-
-    dirs = glob.glob(root_path + "/*/stats.csv")
-
-    values = pd.DataFrame(index =[range(1, len(dirs)+1)])
-    branches=['branch:']
-
-    for directory in dirs:
-        path = Path(directory)
-        branch_name = path.parts[-2]
-        branches.append(branch_name)
-        csv = pd.read_csv(directory)
-        values[branch_name] = csv['f_score']
-
-    plt.boxplot(values, showfliers = True)
-    plt.xticks(range(0, len(branches)),  branches, rotation=-20)
-    plt.grid(True)
-    plt.show()
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description="Run the performance benchmarking tool against a number of git branches")
